Touchpaint - works now
Fixed the mapping of touch to screen coordinates.

Then hacked up, that if you enter anything in Serial monitor it would toggle on or off debug output of touch data.

Also made color list be a table of colors,  and added the color names.  Now output 12 colors across the top of the display, and when you choose a color it prints out the name of color...