@mjs513  @defragster and my - hopefully fix rotation/color issues
Combined work!

To hopefully figure out MADCTL values for what appears to be two versions of the same KeDei display.  Such that we can rotate the screen in all 4 orientations and have things setup correctly.

We are now using an init sequence which is pretty close to what the RPI driver does.

We figured out which version of the displays we have (may be others?), by asking the XPT2046 on the board for it's value associated with one of it's pins, which on these boards is tied to two resistors (R1 and R2), which appear to be setup for a potential resistor divider setup, with
R1 connected to GND
R2 connecto to 3.3v

So some boards have R1 soldered on, and the Analog value will be near 0
Other boards will have R2 soldered on and analog value near 0x7fff

Updated touchpaint and replaced Pink with Navy as Pink is a duplicate color.

Updated ColorPosition test for how it can get and display current MadCTL values